fixed missing arguments in readme examples
The calls to 'draw-solid were missing their 'g2d arguments, so I altered each s-exp to an anon fn so 
'->img would thread them properly.

Changed 'mexican-hat to 'sombrero, which is more precise/concise. fwiw, 'sombrero technically means "hat with brim" but it's still a more precise shorthand for 'sombrero-charro which is what it's called in Mexico.

If you have any feedback, or I shouldn't have changed those fns like that, please let me know. I feel comfortable with clojure, but I still have so much to learn.

By the way, this library is wonderful. I've been having a lot of fun with it so far, and I really appreciate the work you've put into it. If you need any help with anything, I'd love to pitch in.